Rename root package to jsonstore
