+++ tests/test_arraystore_view.py
@@ -3,8 +3,8 @@ import sys
 import sqlite3
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-from sqlite_store.arraystore.table import create_array_table, insert_array
-from sqlite_store.arraystore.view import create_element_concat_view
+from jsonstore.arraystore.table import create_array_table, insert_array
+from jsonstore.arraystore.view import create_element_concat_view
 
 
 def test_element_concat_view_default():
