+++ tests/test_objectstore_fts.py
@@ -3,9 +3,9 @@ import sys
 import sqlite3
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-from sqlite_store.objectstore.table import create_object_table, insert_object
-from sqlite_store.objectstore.view import create_property_concat_view
-from sqlite_store.objectstore.fts import create_property_concat_fts
+from jsonstore.objectstore.table import create_object_table, insert_object
+from jsonstore.objectstore.view import create_property_concat_view
+from jsonstore.objectstore.fts import create_property_concat_fts
 
 
 def test_property_concat_fts_search():
